refactor: moved out information rules converter
